refactor: 디테일 하단 버튼 리팩토링
