--- src/screens/screening/detail/components/bottomDetailButton/index.ts
@@ -1,2 +1,0 @@
-export {default as BottomDetailButton} from './bottomDetailButton';
-export {default as BottomOptionDetailButton} from './bottomOptionDetailButton';
